requirements.txt
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
idna==2.5
nmap==0.0.1
pyaes==1.6.0
Pygments==2.2.0
pyTelegramBotAPI==3.1.0
python-telegram-bot==6.1.0
requests==2.18.1
six==1.10.0
structlog==17.2.0
structlog-pretty==0.1.1
Telethon==0.11.5
urllib3==1.21.1
virtualenv==15.1.0
